Fixed latitude and longitude, both are now doubles

+++ silo-server/src/main/java/pt/tecnico/sauron/silo/domain/Camera.java
@@ -3,9 +3,9 @@ package pt.tecnico.sauron.silo.domain;
 public class Camera {
 
     private String name;
-    private float latitude, longitude;
+    private double latitude, longitude;
 
-    public Camera(String name, float latitude, float longitude) {
+    public Camera(String name, double latitude, double longitude) {
         this.name = name;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -19,19 +19,19 @@ public class Camera {
         this.name = name;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return this.latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return this.longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
